table fixes

+++ UniBook/src/model/Docente.java
@@ -6,7 +6,7 @@ public class Docente extends Utente {
 
 
 	public Docente(String matricola, String nome, String cognome, Date dataNascita, String codicefiscale, String email,
-			String password,int corsoDiLaurea) {
+			String password,Long corsoDiLaurea) {
 		super(matricola, nome, cognome, dataNascita, codicefiscale, email, password,corsoDiLaurea);
 	}
